Plot barplot for total world collagen
I used patchesRun1Both.csv, PatchesRun1Pirf.csv, PatchesRun2Both.csv to run this code. This is to produce the figure 3 in the overleaf document. It is partially complete. I will update it once I have all the components. Both-- both drugs used in the run and Pirf -- only pirf drug is used